Created components package and moved any files
