+++ app/src/main/java/com/numero/jetpack_compose_example/ui/HomeScreen.kt
@@ -5,6 +5,8 @@ import androidx.compose.unaryPlus
 import androidx.ui.material.MaterialTheme
 import androidx.ui.material.surface.Surface
 import com.numero.jetpack_compose_example.core.AppTheme
+import com.numero.jetpack_compose_example.ui.components.ComponentDetailScreen
+import com.numero.jetpack_compose_example.ui.components.ComponentScreen
 
 @Composable
 fun HomeScreen() {
@@ -12,7 +14,9 @@ fun HomeScreen() {
         Surface(color = (+MaterialTheme.colors()).background) {
             when (val screen = ScreenStatus.currentScreen) {
                 is Screen.Home -> ComponentScreen()
-                is Screen.ComponentDetail -> ComponentDetailScreen(screen.component)
+                is Screen.ComponentDetail -> ComponentDetailScreen(
+                    screen.component
+                )
             }
         }
     }
