+++ app/src/main/java/com/numero/jetpack_compose_example/ui/components/ComponentListScreen.kt
@@ -1,4 +1,4 @@
-package com.numero.jetpack_compose_example.ui
+package com.numero.jetpack_compose_example.ui.components
 
 import androidx.compose.Composable
 import androidx.ui.core.Text
@@ -8,6 +8,8 @@ import androidx.ui.material.ListItem
 import androidx.ui.material.TopAppBar
 import com.numero.jetpack_compose_example.core.AppBarLayout
 import com.numero.jetpack_compose_example.model.Component
+import com.numero.jetpack_compose_example.ui.Screen
+import com.numero.jetpack_compose_example.ui.navigateTo
 
 @Composable
 fun ComponentScreen() {
@@ -32,7 +34,11 @@ fun ComponentList() {
         Column {
             componentList.forEach { component ->
                 ListItem(text = component.label) {
-                    navigateTo(Screen.ComponentDetail(component))
+                    navigateTo(
+                        Screen.ComponentDetail(
+                            component
+                        )
+                    )
                 }
             }
         }
